Moved pause, pName below ProductGUIImg, a few cleanups and optimizations.
Occasionally get the error where the score goes up and doesn't stop rising on TimeIsUp(). Haven't seen it since the optimizations, but please keep an eye out for it! As it is a bug we have faced many times already, I am declaring an actual name for it:

BugName: RunOff